@@ -1,6 +1,6 @@
 import React from "react";
 import { useDispatch } from "react-redux";
-import { useHistory } from "react-router-dom";
+import { useNavigate } from "react-router-dom";
 
 import { submitAnalysisWorkflowRun } from "@/modules/wes/actions";
 
@@ -9,13 +9,13 @@ import RunSetupConfirmDisplay from "./RunSetupConfirmDisplay";
 
 const ManagerAnalysisContent = () => {
     const dispatch = useDispatch();
-    const history = useHistory();
+    const navigate = useNavigate();
 
     return <RunSetupWizard
         workflowType="analysis"
         confirmDisplay={(props) => <RunSetupConfirmDisplay runButtonText="Run Analysis" {...props} />}
-        onSubmit={({selectedWorkflow, inputs}) => {
-            dispatch(submitAnalysisWorkflowRun(selectedWorkflow, inputs, "/admin/data/manager/runs", history));
+        onSubmit={({ selectedWorkflow, inputs }) => {
+            dispatch(submitAnalysisWorkflowRun(selectedWorkflow, inputs, "/data/manager/runs", navigate));
         }}
     />;
 };
