@@ -1,6 +1,6 @@
 import React from "react";
 import { useDispatch } from "react-redux";
-import { useHistory } from "react-router-dom";
+import { useNavigate } from "react-router-dom";
 
 import { submitExportWorkflowRun } from "@/modules/wes/actions";
 
@@ -9,13 +9,13 @@ import RunSetupConfirmDisplay from "./RunSetupConfirmDisplay";
 
 const ManagerExportContent = () => {
     const dispatch = useDispatch();
-    const history = useHistory();
+    const navigate = useNavigate();
 
     return <RunSetupWizard
         workflowType="export"
         confirmDisplay={(props) => <RunSetupConfirmDisplay runButtonText="Run Export" {...props} />}
-        onSubmit={({selectedWorkflow, inputs}) => {
-            dispatch(submitExportWorkflowRun(selectedWorkflow, inputs, "/admin/data/manager/runs", history));
+        onSubmit={({ selectedWorkflow, inputs }) => {
+            dispatch(submitExportWorkflowRun(selectedWorkflow, inputs, "/data/manager/runs", navigate));
         }}
     />;
 };
