@@ -1,25 +1,25 @@
-import React, {useCallback, useMemo} from "react";
-import {useDispatch, useSelector} from "react-redux";
-import {useHistory} from "react-router-dom";
+import React, { useCallback, useMemo } from "react";
+import { useDispatch, useSelector } from "react-redux";
+import { useNavigate } from "react-router-dom";
 
-import {Button, Divider, Drawer} from "antd";
+import { Button, Divider, Drawer } from "antd";
 
 import NotificationList from "./NotificationList";
 
-import {hideNotificationDrawer, markAllNotificationsAsRead} from "../../modules/notifications/actions";
+import { hideNotificationDrawer, markAllNotificationsAsRead } from "@/modules/notifications/actions";
 
 
 const NotificationDrawer = React.memo(() => {
     const dispatch = useDispatch();
-    const history = useHistory();
+    const navigate = useNavigate();
 
     const markAllAsRead = useCallback(() => {
         dispatch(markAllNotificationsAsRead()).catch((err) => console.error(err));
     }, [dispatch]);
     const seeAllNotifications = useCallback(() => {
         dispatch(hideNotificationDrawer());
-        history.push("/notifications");
-    }, [dispatch, history]);
+        navigate("/notifications");
+    }, [dispatch, navigate]);
     const hideNotificationDrawer_ = useCallback(() => {
         dispatch(hideNotificationDrawer());
     }, [dispatch]);
@@ -30,17 +30,17 @@ const NotificationDrawer = React.memo(() => {
 
     const notificationDrawerVisible = useSelector((state) => state.notifications.drawerVisible);
 
-    return <Drawer bodyStyle={{padding: 0}} title="Unread Notifications"
+    return <Drawer bodyStyle={{ padding: 0 }} title="Unread Notifications"
                    open={notificationDrawerVisible}
                    width={500}
                    onClose={hideNotificationDrawer_}>
-        <div style={{padding: "16px 24px", display: "flex", gap: "16px"}}>
-            <Button style={{flex: 1}} onClick={markAllAsRead}
+        <div style={{ padding: "16px 24px", display: "flex", gap: "16px" }}>
+            <Button style={{ flex: 1 }} onClick={markAllAsRead}
                     disabled={!unreadNotifications} loading={isMarkingAllAsRead}>Mark All as Read</Button>
-            <Button style={{flex: 1}} onClick={seeAllNotifications}>See All Notifications</Button>
+            <Button style={{ flex: 1 }} onClick={seeAllNotifications}>See All Notifications</Button>
         </div>
-        <Divider style={{margin: 0}} />
-        <div style={{padding: "0 24px"}}>
+        <Divider style={{ margin: 0 }} />
+        <div style={{ padding: "0 24px" }}>
             <NotificationList small={true} notifications={unreadNotifications} />
         </div>
     </Drawer>;
